Move auth to features
