+++ src/containers/appRoot.tsx
@@ -3,9 +3,9 @@ import { Route, Switch, Redirect, NavLink } from 'react-router-dom';
 import { TimingsPageContainer } from './timingList';
 import { connect } from 'react-redux';
 import { AppState } from '../store';
-import { getSession } from '../store/auth/actions';
+import { getSession } from '../features/auth/state/actions';
 import { assertNever } from '../shared';
-import { LoginPageContainer, LogoutPageContainer } from './auth';
+import { LoginPageContainer, LogoutPageContainer } from '../features/auth/containers/auth';
 import { Navbar } from '../components/navbar';
 import { Logo } from '../components/logo';
 import { UserSession } from '../shared/types';
