+++ src/features/auth/api/getSession/index.ts
@@ -1,6 +1,6 @@
+import { ensureSuccessStatusCode, readJson } from 'shared/http';
+import { UserSession } from 'shared/types';
 import { metadata } from './contract';
-import { ensureSuccessStatusCode, readJson } from '../../shared/http';
-import { UserSession } from '../../shared/types';
 
 export async function getSession(): Promise<UserSession> {
     const response = await fetch(metadata.url, { method: metadata.method });
