+++ src/features/auth/api/logout/contract.ts
@@ -1,4 +1,4 @@
-import { RequestMetadata } from '../../shared/http';
+import { RequestMetadata } from 'shared/http';
 
 export const metadata: RequestMetadata = {
     method: 'POST',
