+++ src/features/auth/api/logout/index.ts
@@ -1,5 +1,5 @@
+import { ensureSuccessStatusCode } from 'shared/http';
 import { metadata } from './contract';
-import { ensureSuccessStatusCode } from '../../shared/http';
 
 export async function logout() {
     const response = await fetch(metadata.url, { method: metadata.method });
