+++ tsconfig.json
@@ -2,6 +2,7 @@
     "compilerOptions": {
         "outDir": "dist",
         "rootDir": "src",
+        "baseUrl": "src",
 
         "module": "commonjs",
         "esModuleInterop": true,
@@ -14,7 +15,7 @@
         "strictNullChecks": true,
         "allowSyntheticDefaultImports": true,
         "lib": ["es2017", "dom"],
-        
+
         "jsx": "react",
     },
     "include": ["src"],
